+++ .eslintignore
@@ -1,0 +1,3 @@
+/dist
+/front
+/node_modules
