+++ .gitignore
@@ -14,3 +14,5 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+/front
