+++ .prettierrc.json
@@ -1,0 +1,9 @@
+{
+  "endOfLine": "auto",
+  "semi": true,
+  "tabWidth": 2,
+  "singleQuote": true,
+  "trailingComma": "all",
+  "arrowParens": "always",
+  "printWidth": 80
+}
