+++ nodemon.json
@@ -1,0 +1,11 @@
+{
+  "watch": ["src"],
+  "ext": "ts,js,json",
+  "ignore": ["node_modules", "front", "dist"],
+  "exec": "node --loader ts-node/esm ./src/index.ts",
+  "execDefault": "ts-node -r tsconfig-paths/register ./src/index.ts",
+  "restartable": "rs",
+  "env": {
+    "NODE_ENV": "development"
+  }
+}
